Add explicit double cast for health in TR5

+++ TR-SaveMaster/TRCUtilities.cs
@@ -572,7 +572,7 @@ namespace TR_SaveMaster
 
             if (trbHealth.Enabled)
             {
-                WriteHealthValue(trbHealth.Value);
+                WriteHealthValue((double)trbHealth.Value);
             }
         }
